requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
apscheduler>=3.10
python-dotenv>=1.0
requests>=2.31
